Unifiy Sarama config building
