+++ pkg/channel/distributed/common/kafka/admin/admin_kafka.go
@@ -86,8 +86,20 @@ func NewKafkaAdminClient(ctx context.Context, saramaConfig *sarama.Config, clien
 	username := string(kafkaSecret.Data[constants.KafkaSecretKeyUsername])
 	password := string(kafkaSecret.Data[constants.KafkaSecretKeyPassword])
 
+	client.InitSaramaConfig(saramaConfig)
+	saramaConfig.ClientID = clientId
+
 	// Update The Sarama ClusterAdmin Configuration With Our Values
-	client.UpdateSaramaConfig(saramaConfig, clientId, username, password)
+	err = client.UpdateSaramaConfigWithKafkaAuthConfig(saramaConfig, &client.KafkaAuthConfig{
+		SASL: &client.KafkaSaslConfig{
+			User:     username,
+			Password: password,
+		},
+	})
+	if err != nil {
+		logger.Fatal("Unable to set SASL username and password on Sarama config", zap.Error(err))
+		return nil, err
+	}
 
 	// Create A New Sarama ClusterAdmin
 	clusterAdmin, err := NewClusterAdminWrapper(brokers, saramaConfig)
